feat: separated rootConfig
